// employee/src/leave/leave-request/enums/leave-request-duration-type.enum.ts
export enum LeaveRequestDurationTypeEnEnum {
  DATE_RANGE = 'DATE_RANGE',
  FIRST_HALF_DAY = 'FIRST_HALF_DAY',
  SECOND_HALF_DAY = 'SECOND_HALF_DAY'
}

export enum LeaveRequestDurationTypeKhEnum {
  DATE_RANGE = 'ចន្លេាះពេល',
  FIRST_HALF_DAY = 'ពាក់កណ្តាលថ្ងៃទីមួយ',
  SECOND_HALF_DAY = 'ពាក់កណ្តាលថ្ងៃទីពីរ'
}